requirements.txt
streamlit
pandas
numpy
matplotlib
seaborn
nltk
wordcloud
pillow
os
plotly
